Completely remove followed streams that appear in the authenticated list.
This was causing an anomaly with JAObjectListView, since there was no
way we could retrieve an object in a specific section (even though the
one in the featured list didn’t show, it still existed in the array).
Deleting an object would delete both view items if they existed in both.